optimize include blobstorage_pdisk.h
