+++ ydb/core/blobstorage/vdisk/common/blobstorage_dblogcutter.cpp
@@ -6,6 +6,7 @@
 #include "vdisk_lsnmngr.h"
 #include <ydb/core/base/appdata.h>
 #include <ydb/core/blobstorage/base/utility.h>
+#include <ydb/core/blobstorage/pdisk/blobstorage_pdisk.h>
 
 #include <library/cpp/monlib/service/pages/templates.h>
