+++ ydb/core/blobstorage/vdisk/common/vdisk_context.h
@@ -12,7 +12,7 @@
 #include <ydb/core/base/blobstorage_common.h>
 #include <ydb/core/blobstorage/base/ptr.h>
 #include <ydb/core/blobstorage/groupinfo/blobstorage_groupinfo.h>
-#include <ydb/core/blobstorage/pdisk/blobstorage_pdisk.h>
+// #include <ydb/core/blobstorage/pdisk/blobstorage_pdisk.h>
 
 namespace NKikimr {
     class TCostModel;
@@ -146,8 +146,8 @@ namespace NKikimr {
             }
         }
 
-        template <class TActorSystemOrCtx>
-        bool CheckPDiskResponseReadable(const TActorSystemOrCtx &actorSystemOrCtx, const NPDisk::TEvChunkReadResult &ev, const TString &message = {}) {
+        template <class TActorSystemOrCtx, class TEvChunkReadResult>
+        bool CheckPDiskResponseReadable(const TActorSystemOrCtx &actorSystemOrCtx, const TEvChunkReadResult &ev, const TString &message = {}) {
             if (!ev.Data.IsReadable()) {
                 LOG_ERROR(actorSystemOrCtx, NKikimrServices::BS_VDISK_OTHER,
                         VDISKP(VDiskLogPrefix,
