+++ ydb/core/blobstorage/vdisk/hullop/hullcompdelete/blobstorage_hullcompdelete.cpp
@@ -1,6 +1,37 @@
 #include "blobstorage_hullcompdelete.h"
 
+#include <ydb/core/blobstorage/pdisk/blobstorage_pdisk.h>
+
 namespace NKikimr {
+
+    void TDelayedCompactionDeleterInfo::ProcessReleaseQueue(const TActorContext& ctx, const TActorId& hugeKeeperId, const TActorId& skeletonId,
+                const TPDiskCtxPtr& pdiskCtx, const TVDiskContextPtr& vctx) {
+            // if we have no snapshots, we can safely process all messages; otherwise we can process only those messages
+            // which do not have snapshots created before the point of compaction
+            while (ReleaseQueue) {
+                TReleaseQueueItem& item = ReleaseQueue.front();
+                if (CurrentSnapshots.empty() || (item.RecordLsn <= CurrentSnapshots.begin()->first)) {
+                    // matching record -- commit it to huge hull keeper and throw out of the queue
+                    if (item.WId) {
+                        ctx.Send(hugeKeeperId, new TEvHullFreeHugeSlots(std::move(item.RemovedHugeBlobs),
+                            item.RecordLsn, item.Signature, item.WId));
+                    } else {
+                        Y_ABORT_UNLESS(item.RemovedHugeBlobs.Empty());
+                    }
+                    if (item.ChunksToForget) {
+                        LOG_DEBUG(ctx, NKikimrServices::BS_VDISK_CHUNKS, VDISKP(vctx->VDiskLogPrefix,
+                            "FORGET: PDiskId# %s ChunksToForget# %s", pdiskCtx->PDiskIdString.data(),
+                            FormatList(item.ChunksToForget).data()));
+                        TActivationContext::Send(new IEventHandle(pdiskCtx->PDiskId, skeletonId, new NPDisk::TEvChunkForget(
+                            pdiskCtx->Dsk->Owner, pdiskCtx->Dsk->OwnerRound, std::move(item.ChunksToForget))));
+                    }
+                    ReleaseQueue.pop_front();
+                } else {
+                    // we have no matching record
+                    break;
+                }
+            }
+        }
 
     class TDelayedCompactionDeleterActor : public TActor<TDelayedCompactionDeleterActor> {
         // pointer to database general data; actually we need only HugeKeeperID from that data
