+++ ydb/core/blobstorage/vdisk/skeleton/skeleton_oos_tracker.cpp
@@ -5,6 +5,7 @@
 #include <ydb/core/blobstorage/vdisk/common/vdisk_mongroups.h>
 #include <ydb/core/blobstorage/vdisk/common/vdisk_mon.h>
 #include <ydb/core/blobstorage/vdisk/common/vdisk_pdiskctx.h>
+#include <ydb/core/blobstorage/pdisk/blobstorage_pdisk.h>
 #include <library/cpp/monlib/service/pages/templates.h>
 
 namespace NKikimr {
