+++ ydb/core/blobstorage/vdisk/synclog/blobstorage_synclogwriteparts.h
@@ -2,6 +2,7 @@
 
 #include "defs.h"
 #include <ydb/core/blobstorage/vdisk/common/vdisk_pdiskctx.h>
+#include <ydb/core/blobstorage/pdisk/blobstorage_pdisk.h>
 
 namespace NKikimr {
     namespace NSyncLog {
